[REF] - S#13590 - Move coop_membership_access_rights to coop_membership
