+++ intercoop_addons/coop_membership/demo/res_groups.xml
