+++ intercoop_addons/coop_membership/models/__init__.py
@@ -14,3 +14,4 @@ from . import shift_registration
 from . import shift_extension
 from . import shift_extension_type
 from . import res_config
+from . import res_users
