+++ intercoop_addons/coop_membership/models/res_partner.py
@@ -126,6 +126,10 @@ class ResPartner(models.Model):
         compute="_compute_number_of_associated_people",
         store=True)
 
+    related_user_id = fields.Many2one('res.users',
+                                      compute="_compute_related_user",
+                                      string="Related User")
+
     # Constraint Section
     @api.multi
     @api.constrains('is_member',
@@ -459,3 +463,42 @@ class ResPartner(models.Model):
             next_shift_date = start_date_object_tz.strftime('%Y-%m-%d')
 
         return next_shift_time, next_shift_date
+
+    @api.multi
+    def _compute_related_user(self):
+        """
+        Function to compute the related user of the partner
+        """
+        res_user_env = self.env["res.users"]
+        for partner in self:
+            related_users = res_user_env.search(
+                [('partner_id', '=', partner.id)])
+            partner.related_user_id = \
+                related_users and related_users[0] or False
+
+    @api.multi
+    def action_create_new_user(self):
+        """
+        Function to activate the User Creation Form
+        """
+        self.ensure_one()
+
+        # Prepare context for default value
+        context = self.env.context.copy()
+        context.update({
+            'default_partner_id': self.id,
+            'default_name': self.name,
+            'default_login': self.email,
+            'default_email': self.email
+        })
+
+        return {
+            'name': _('Create New User'),
+            'context': context,
+            'view_type': 'form',
+            'view_mode': 'tree',
+            'res_model': 'res.users',
+            'views': [(self.env.ref("base.view_users_form").id, 'form')],
+            'type': 'ir.actions.act_window',
+        }
+
