+++ intercoop_addons/coop_membership/models/shift_registration.py
@@ -36,6 +36,10 @@ class ShiftRegistration(models.Model):
     related_extension_id = fields.Many2one('shift.extension',
                                            string="Related Shift Extensions")
 
+    related_shift_state = fields.Selection(related="shift_id.state",
+                                           store=False,
+                                           string="Shift State")
+
     @api.model
     def create(self, vals):
         partner_id = vals.get('partner_id', False)
@@ -45,17 +49,21 @@ class ShiftRegistration(models.Model):
             raise UserError(_(
                 """You can't register this partner on a shift because """
                 """he isn't registered on a template"""))
-        return super(ShiftRegistration, self).create(vals)
+        res = super(ShiftRegistration, self).create(vals)
+        # Do not allow member with Up to date status register make up
+        # in a ABCD shift on a ABCD tickets
+        res.checking_shift_attendance()
+        return res
 
     @api.multi
     def action_create_extension(self):
-        '''
+        """
         @Function triggered by a button on Attendance tree view
         to create extension automatically for a member:
             - Extension Type: Extension
             - Start Date: registration start date
             - End Date: Next Shift Date
-        '''
+        """
         shift_extension_env = self.env['shift.extension']
         for registration in self:
             partner = registration.partner_id
@@ -134,7 +142,7 @@ class ShiftRegistration(models.Model):
 
     @api.multi
     def write(self, vals):
-        '''
+        """
         Overide write function to update point counter for member
             + Standard:
                 Add 1: Status is Attended / Replaced and template not created
@@ -144,7 +152,7 @@ class ShiftRegistration(models.Model):
                 Add 1: Status is Attended / Replaced
                 Deduct 1: Status is Excused / Waiting and template created
                 Deduct 1: Status is Absent
-        '''
+        """
         point_counter_env = self.env['shift.counter.event']
         vals_state = vals.get('state')
         for shift_reg in self:
@@ -203,4 +211,52 @@ class ShiftRegistration(models.Model):
                     point_counter_env.sudo().with_context(
                         automatic=True).create(counter_vals)
 
-        return super(ShiftRegistration, self).write(vals)
+        res = super(ShiftRegistration, self).write(vals)
+        if 'template_created' in vals or 'shift_ticket_id' in vals:
+            self.checking_shift_attendance()
+        return res
+
+    @api.multi
+    @api.onchange("shift_id")
+    def onchange_shift_id(self):
+        # Use the context value for default
+        is_standard_ticket = self.env.context.get("is_standard_ticket", False)
+        ticket_type_product = False
+        if is_standard_ticket:
+            ticket_type_product = \
+                self.env.ref("coop_shift.product_product_shift_standard")
+        else:
+            ticket_type_product = \
+                self.env.ref("coop_shift.product_product_shift_ftop")
+        for reg in self:
+            reg.shift_ticket_id = reg.shift_id.shift_ticket_ids.filtered(
+                lambda t: t.product_id == ticket_type_product)
+
+    @api.multi
+    def checking_shift_attendance(self):
+        """
+        @Function to check the attendance:
+            - Do not allow member with Up to date status registers make up
+            in a ABCD shift on a ABCD tickets
+        """
+        ignore_checking = \
+            self.env.context.get('ignore_checking_attendance', False)
+        if ignore_checking:
+            return
+        uptodate_list = []
+        for shift_reg in self:
+            shift_type = shift_reg.shift_id and \
+                shift_reg.shift_id.shift_type_id
+            if shift_type and not shift_type.is_ftop and \
+                    shift_reg.shift_type == 'standard' and \
+                    not shift_reg.template_created and \
+                    shift_reg.state not in ['replacing', 'replaced'] and \
+                    shift_reg.partner_id.working_state == 'up_to_date':
+                uptodate_list.append('- [%s] %s' % (
+                    shift_reg.partner_id.barcode_base,
+                    shift_reg.partner_id.name))
+        if uptodate_list:
+            raise UserError(_("Warning! You cannot enter make-ups of " +
+                              "the following members " +
+                              "as they are up-to-date: \n\n%s") %
+                            '\n'.join(uptodate_list))
