+++ intercoop_addons/coop_membership/security/ir_module_category.xml
