+++ intercoop_addons/coop_membership/security/res_group.xml
@@ -55,13 +55,13 @@
     <record model="res.groups" id="subscriptions_creator">
         <field name="name">Create Subscriptions</field>
         <field name="category_id" ref="category_subscriptions" />
-        <field name="implied_ids" eval="[(4, ref('coop_membership_access_rights.subscriptions_reader'))]" />
+        <field name="implied_ids" eval="[(4, ref('coop_membership.subscriptions_reader'))]" />
     </record>
 
     <record model="res.groups" id="subscriptions_creator_payment">
         <field name="name">Create Subscriptions and payment</field>
         <field name="category_id" ref="category_subscriptions" />
-        <field name="implied_ids" eval="[(4, ref('coop_membership_access_rights.subscriptions_creator'))]" />
+        <field name="implied_ids" eval="[(4, ref('coop_membership.subscriptions_creator'))]" />
     </record>
 
     <record model="res.groups" id="subscriptions_can_change_fundraising_category">
@@ -86,8 +86,9 @@
         <field name="category_id" ref="coop_shift.module_category_shift_management" />
         <field name="users" eval="[(4, ref('base.user_root'))]"/>
     </record>
+
     <record id="base.group_user" model="res.groups">
-        <field name="implied_ids" eval="[(4, ref('group_membership_see_associated_people'))]"></field>
+        <field name="implied_ids" eval="[(4, ref('group_membership_see_associated_people'))]"/>
 	</record>
 
 </odoo>
