+++ intercoop_addons/coop_membership/views/res_users_view.xml
