+++ intercoop_addons/coop_membership/wizard/capital_fundraising_wizard.py
@@ -3,7 +3,7 @@
 # @author: Sylvain LE GAL (https://twitter.com/legalsylvain)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from openerp import models, fields, api
+from openerp import models, fields
 
 
 class CapitalFundraisingWizard(models.TransientModel):
@@ -11,8 +11,7 @@ class CapitalFundraisingWizard(models.TransientModel):
 
     def default_can_change_fundraising_category(self):
         return self.user_has_groups(
-            'coop_membership_access_rights.' +
-            'subscriptions_can_change_fundraising_category')
+            'coop_membership.subscriptions_can_change_fundraising_category')
 
     can_change_fundraising_category = fields.Boolean(
         string="Can Change fundraising Category",
