+++ intercoop_addons/coop_membership/wizard/view_capital_fundraising_wizard.xml
