--- louve_addons/coop_membership_access_rights/README.rst
@@ -1,18 +1,0 @@
-.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
-    :alt: License: AGPL-3
-
-Coop Membership Access Rights
-================
-
-Handle the views and rights to help workers to manage members
-
-* Merge all res.partner views into a single one
-* Create specific user rights for member who must do only limited actions
-
-Credits
-=======
-
-Contributors
-------------
-
-* Julien Weste (julien.weste@akretion.com)
