--- louve_addons/coop_membership_access_rights/models/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from . import res_users
-from . import res_partner
-from . import shift_shift
-from . import shift_registration
