--- louve_addons/coop_membership_access_rights/views/action.xml
@@ -1,117 +1,0 @@
-<?xml version="1.0"?>
-<!-- Copyright (C) 2016-Today: La Louve (<http://www.lalouve.net/>)
-	@author: La Louve
-	License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html). -->
-
-<odoo>
-	<!-- Member -->
-    <record id="bdm_action1" model="ir.actions.act_window.view">
-       <field eval="1" name="sequence"/>
-       <field name="view_mode">tree</field>
-       <field name="view_id" ref="coop_membership_access_rights.view_bdm_tree_no_create"/>
-       <field name="act_window_id" ref="action_coop_membership_access_rights"/>
-    </record>
-    <record id="bdm_action2" model="ir.actions.act_window.view">
-       <field eval="2" name="sequence"/>
-       <field name="view_mode">form</field>
-       <field name="view_id" ref="coop_membership_access_rights.view_partner_form_no_create"/>
-       <field name="act_window_id" ref="action_coop_membership_access_rights"/>
-    </record>
-
-	<!-- Associated People : New form and tree view for new group 'group_membership_access_user' -->
-	<record id="bdm_action1_associated_people" model="ir.actions.act_window.view">
-		<field eval="1" name="sequence" />
-		<field name="view_mode">tree</field>
-		<field name="view_id" ref="coop_membership_access_rights.view_bdm_tree_no_create" />
-		<field name="act_window_id" ref="coop_membership.action_res_partner_associated" />
-	</record>
-	<record id="bdm_action2_associated_people" model="ir.actions.act_window.view">
-		<field eval="2" name="sequence" />
-		<field name="view_mode">form</field>
-		<field name="view_id" ref="coop_membership_access_rights.view_partner_form_no_create" />
-		<field name="act_window_id" ref="coop_membership.action_res_partner_associated" />
-	</record>
-
-	<!-- Interested People : New form and tree view for new group 'group_membership_access_user' -->
-	<record id="bdm_action1_interested_people" model="ir.actions.act_window.view">
-		<field eval="1" name="sequence" />
-		<field name="view_mode">tree</field>
-		<field name="view_id" ref="coop_membership_access_rights.view_bdm_tree_no_create" />
-		<field name="act_window_id"
-			ref="coop_membership.action_res_partner_interested_people" />
-	</record>
-	<record id="bdm_action2_interested_people" model="ir.actions.act_window.view">
-		<field eval="2" name="sequence" />
-		<field name="view_mode">form</field>
-		<field name="view_id" ref="coop_membership_access_rights.view_partner_form_no_create" />
-		<field name="act_window_id"
-			ref="coop_membership.action_res_partner_interested_people" />
-	</record>
-
-	<!-- Former Members : New form and tree view for new group 'group_membership_access_user' -->
-	<record id="bdm_action1_former_members" model="ir.actions.act_window.view">
-		<field eval="1" name="sequence" />
-		<field name="view_mode">tree</field>
-		<field name="view_id" ref="coop_membership_access_rights.view_bdm_tree_no_create" />
-		<field name="act_window_id" ref="coop_membership.action_res_partner_former_member" />
-	</record>
-	<record id="bdm_action2_former_members" model="ir.actions.act_window.view">
-		<field eval="2" name="sequence" />
-		<field name="view_mode">form</field>
-		<field name="view_id" ref="coop_membership_access_rights.view_partner_form_no_create" />
-		<field name="act_window_id" ref="coop_membership.action_res_partner_former_member" />
-	</record>
-
-    <!-- Former Associated People : New form and tree view for new group 'group_membership_access_user' -->
-    <record id="bdm_action1_former_associated_people" model="ir.actions.act_window.view">
-        <field eval="1" name="sequence" />
-        <field name="view_mode">tree</field>
-        <field name="view_id" ref="coop_membership_access_rights.view_bdm_tree_no_create" />
-        <field name="act_window_id" ref="coop_membership.action_res_partner_former_associated_people" />
-    </record>
-    <record id="bdm_action2_former_associated_people" model="ir.actions.act_window.view">
-        <field eval="2" name="sequence" />
-        <field name="view_mode">form</field>
-        <field name="view_id" ref="coop_membership_access_rights.view_partner_form_no_create" />
-        <field name="act_window_id" ref="coop_membership.action_res_partner_former_associated_people" />
-    </record>
-
-    <!-- View actions for shift attendance -->
-    <record id="view_action_tree_shift_attendance_entry" model="ir.actions.act_window.view">
-        <field eval="1" name="sequence" />
-        <field name="view_mode">tree</field>
-        <field name="view_id" ref="coop_shift.view_shift_tree" />
-        <field name="act_window_id" ref="coop_membership.action_shift_attendance_entry_view" />
-    </record>
-
-    <record id="view_action_form_shift_attendance_entry" model="ir.actions.act_window.view">
-        <field eval="2" name="sequence" />
-        <field name="view_mode">form</field>
-        <field name="view_id" ref="coop_membership.view_shift_form_attendance_entry" />
-        <field name="act_window_id" ref="coop_membership.action_shift_attendance_entry_view" />
-    </record>
-
-  <record id="action_badge_to_print_form" model="ir.actions.act_window.view">
-       <field eval="32" name="sequence"/>
-       <field name="view_mode">form</field>
-       <field name="view_id" ref="coop_membership_access_rights.view_partner_form_no_create"/>
-       <field name="act_window_id" ref="coop_print_badge.action_badge_to_print"/>
-    </record>
-
-    <record id="coop_membership.action_coop_all_partner_view" model="ir.actions.act_window">
-      <field name="search_view_id" ref="view_bdm_search"/>
-    </record>
-    <record id="view_all_partner_tree_action" model="ir.actions.act_window.view">
-       <field eval="1" name="sequence"/>
-       <field name="view_mode">tree</field>
-       <field name="view_id" ref="coop_membership_access_rights.view_res_partner_all_members_tree"/>
-       <field name="act_window_id" ref="coop_membership.action_coop_all_partner_view"/>
-    </record>
-    <record id="view_all_member_tree_action" model="ir.actions.act_window.view">
-       <field eval="2" name="sequence"/>
-       <field name="view_mode">form</field>
-       <field name="view_id" ref="coop_membership_access_rights.view_partner_form"/>
-       <field name="act_window_id" ref="coop_membership.action_coop_all_partner_view"/>
-    </record>
-
-</odoo>
