--- louve_addons/coop_membership_access_rights/views/menu.xml
@@ -1,84 +1,0 @@
-<?xml version="1.0"?>
-<odoo>
-    <record model="ir.ui.menu" id="coop_shift.shift_main_menu">
-        <field name="groups_id" eval="[(4, ref('coop_membership_access_rights.group_membership_access_user')), (4, ref('coop_membership_access_rights.subscriptions_reader'))]"/>
-    </record>
-
-    <record model="ir.ui.menu" id="coop_membership.menu_res_partner_interested_people">
-        <field name="groups_id" eval="[(4, ref('coop_membership_access_rights.group_membership_access_manager'))]"/>
-    </record>
-
-    <record model="ir.ui.menu" id="coop_membership.menu_res_partner_former_member">
-        <field name="groups_id" eval="[(4, ref('coop_membership_access_rights.group_membership_access_manager'))]"/>
-    </record>
-
-    <record model="ir.ui.menu" id="coop_shift.shift_menu">
-        <field name="groups_id" eval="[(4, ref('coop_membership_access_rights.group_membership_access_edit'))]"/>
-    </record>
-
-    <record model="ir.ui.menu" id="coop_shift.menu_action_registration">
-        <field name="groups_id" eval="[(4, ref('coop_membership_access_rights.group_membership_access_edit'))]"/>
-    </record>
-
-    <record model="ir.ui.menu" id="mass_mailing.mass_mailing_menu_root">
-        <field name="groups_id" eval="[(4, ref('base.group_configuration')), (4, ref('mass_mailing.group_mass_mailing_user'))]"/>
-    </record>
-
-    <record model="ir.ui.menu" id="coop_membership.menu_coop_subscriptions_submain">
-        <field name="groups_id" eval="[(4, ref('coop_membership_access_rights.subscriptions_reader'))]"/>
-    </record>
-
-    <record model="ir.ui.menu" id="coop_membership.menu_account_invoice_subscriptions">
-        <field name="groups_id" eval="[(4, ref('coop_membership_access_rights.subscriptions_reader'))]"/>
-    </record>
-
-    <record model="ir.ui.menu" id="coop_membership.menu_capital_fundraising_wizard">
-        <field name="groups_id" eval="[(4, ref('coop_membership_access_rights.subscriptions_creator'))]"/>
-    </record>
-
-    <record model="ir.ui.menu" id="coop_membership.menu_account_payment_subscriptions">
-        <field name="groups_id" eval="[(4, ref('coop_membership_access_rights.subscriptions_creator_payment'))]"/>
-    </record>
-
-    <record model="ir.ui.menu" id="capital_subscription.menu_capital_fundraising_wizard">
-        <field name="groups_id" eval="[(4, ref('coop_membership_access_rights.subscriptions_creator'))]" />
-    </record>
-
-    <!-- Interested People : Add group 'group_membership_access_user' -->
-    <record model="ir.ui.menu" id="coop_membership.menu_res_partner_interested_people">
-        <field name="groups_id" eval="[(4, ref('coop_membership_access_rights.group_membership_access_user'))]"/>
-    </record>
-
-    <!-- Former Members : Add group 'group_membership_access_user' -->
-    <record model="ir.ui.menu" id="coop_membership.menu_res_partner_former_member">
-        <field name="groups_id" eval="[(4, ref('coop_membership_access_rights.group_membership_access_user'))]"/>
-    </record>
-
-    <record id="coop_shift.menu_shift_shift" model="ir.ui.menu">
-        <field eval="8" name="sequence"/>
-    </record>
-    <record model="ir.ui.menu" id="coop_print_badge.menu_badge_to_print">
-      <field name="groups_id" eval="[(4, ref('coop_membership_access_rights.group_membership_access_photo'))]"/>
-    </record>
-
-    <menuitem action="coop_membership.action_shift_attendance_entry_view" id="menu_shift_attendance_entry" name="Attendance Entry" parent="coop_shift.shift_menu" sequence="9"/>
-
-    <record id="coop_membership.menu_res_partner_members" model="ir.ui.menu">
-        <field eval="2" name="sequence"/>
-    </record>
-    <record id="coop_membership.menu_res_partner_associated" model="ir.ui.menu">
-        <field eval="3" name="sequence"/>
-        <field name="groups_id" eval="[(4, ref('coop_membership_access_rights.group_membership_see_associated_people'))]"/>
-    </record>
-    <record id="coop_membership.menu_res_partner_former_member" model="ir.ui.menu">
-        <field eval="4" name="sequence"/>
-    </record>
-    <record id="coop_membership.menu_res_partner_interested_people" model="ir.ui.menu">
-        <field eval="5" name="sequence"/>
-    </record>
-    <record id="coop_membership.menu_res_partner_former_associated_people" model="ir.ui.menu">
-        <field eval="6" name="sequence"/>
-        <field name="groups_id" eval="[(4, ref('coop_membership_access_rights.group_membership_see_associated_people'))]"/>
-    </record>
-
-</odoo>
