--- louve_addons/coop_membership_access_rights/views/shift_registration_view.xml
@@ -1,93 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-        <record id="view_shift_registration_tree_confirm_inherit" model="ir.ui.view">
-            <field name="name">shift.registration.tree.confirm.inherit</field>
-            <field name="model">shift.registration</field>
-            <field name="inherit_id" ref="coop_shift.view_shift_registration_tree_confirm"/>
-            <field name="arch" type="xml">
-                <field name="is_related_shift_ftop" position="after">
-                    <field invisible="1" name="related_shift_state"/>
-                </field>
-                <button name="button_reg_close" position="attributes">
-                    <attribute name="attrs">{'invisible': ['|', '|', ('state', '!=', 'open'),('is_related_shift_ftop', '=', True), ('related_shift_state', '!=', 'entry')]}</attribute>
-                </button>
-                <button name="confirm_registration" position="attributes">
-                    <attribute name="attrs">{'invisible': ['|', '|', ('state', '!=', 'draft'),('is_related_shift_ftop', '=', True), ('related_shift_state', '!=', 'entry')]}</attribute>
-                </button>
-                <button name="button_reg_absent" position="attributes">
-                    <attribute name="attrs">{'invisible': ['|', '|', ('state', '!=', 'open'),('is_related_shift_ftop', '=', True), ('related_shift_state', '!=', 'entry')]}</attribute>
-                </button>
-                <button name="button_reg_cancel" position="attributes">
-                    <attribute name="attrs">{'invisible': ['|', '|', ('state', '!=', 'draft'),('is_related_shift_ftop', '=', True), ('related_shift_state', '!=', 'entry')]}</attribute>
-                </button>
-                <button name="button_reg_excused" position="attributes">
-                    <attribute name="attrs">{'invisible': ['|', '|', ('state', '!=', 'open'),('is_related_shift_ftop', '=', True), ('related_shift_state', '!=', 'entry')]}</attribute>
-                </button>
-                <button name="do_draft" position="attributes">
-                    <attribute name="attrs">{'invisible': ['|', '|', ('state', 'not in', ['cancel', 'done', 'absent', 'excused', 'replaced', 'replacing', 'open']),('is_related_shift_ftop', '=', True), ('related_shift_state', '!=', 'entry')]}</attribute>
-                </button>
-                <button name="%(coop_shift.action_replace_registration)d" position="attributes">
-                    <attribute name="attrs">{'invisible': ['|', '|', ('state', '!=', 'open'),('is_related_shift_ftop', '=', True), ('related_shift_state', '!=', 'entry')]}</attribute>
-                </button>
-            </field>
-        </record>
-
-
-        <record id="view_shift_registration_form_inherit" model="ir.ui.view">
-            <field name="name">shift.registration.confirm.inherit</field>
-            <field name="model">shift.registration</field>
-            <field name="inherit_id" ref="coop_shift.view_shift_registration_form"/>
-            <field name="arch" type="xml">
-                <field name="shift_type" position="after">
-                    <field invisible="1" name="related_shift_state"/>
-                </field>
-                <button name="button_reg_close" position="attributes">
-                    <attribute name="states"></attribute>
-                    <attribute name="attrs">{'invisible': ['|', ('related_shift_state', '!=', 'entry'), ('state', '!=', 'open')]}</attribute>
-                </button>
-
-                <button name="button_reg_absent" position="attributes">
-                    <attribute name="states"></attribute>
-                    <attribute name="attrs">{'invisible': ['|', ('related_shift_state', '!=', 'entry'), ('state', '!=', 'open')]}</attribute>
-                </button>
-
-                <button name="button_reg_excused" position="attributes">
-                    <attribute name="states"></attribute>
-                    <attribute name="attrs">{'invisible': ['|', ('related_shift_state', '!=', 'entry'), ('state', '!=', 'open')]}</attribute>
-                </button>
-
-                <button name="%(coop_shift.action_replace_registration)d" position="attributes">
-                    <attribute name="states"></attribute>
-                    <attribute name="attrs">{'invisible': ['|', ('related_shift_state', '!=', 'entry'), ('state', '!=', 'open')]}</attribute>
-                </button>
-
-            </field>
-        </record>
-
-
-        <!-- Views -->
-        <record id="view_shift_registration_tree_access_rights_inherit" model="ir.ui.view">
-            <field name="name">shift.registration.tree.inherit</field>
-            <field name="model">shift.registration</field>
-            <field name="inherit_id" ref="coop_membership.view_shift_registration_tree_inherit"/>
-            <field name="arch" type="xml">
-                <button name="action_create_extension" position="attributes">
-                    <attribute name="groups">coop_membership_access_rights.group_shift_create_extension_from_attendance</attribute>
-                    <attribute name="attrs">{'invisible': ['|', ('related_shift_state', '!=', 'entry'), ('related_extension_id', '!=', False)]}</attribute>
-                </button>
-            </field>
-        </record>
-
-        <record id="view_shift_registration_tree_confirm_one2many_inherit" model="ir.ui.view">
-            <field name="name">shift.registration.tree.confirm.one2many.inherit</field>
-            <field name="model">shift.registration</field>
-            <field name="inherit_id" ref="coop_membership.view_shift_registration_tree_confirm_one2many"/>
-            <field name="arch" type="xml">
-                <button name="action_create_extension" position="attributes">
-                    <attribute name="groups">coop_membership_access_rights.group_shift_create_extension_from_attendance</attribute>
-                </button>
-            </field>
-        </record>
-    </data>
-</openerp>
